+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/PartyFinderEvents.kt
@@ -1,9 +1,9 @@
 package cloud.glitchdev.rfu.events.managers
 
 import cloud.glitchdev.rfu.events.AbstractEventManager
-import cloud.glitchdev.rfu.feature.other.ignore.IgnoreUtils
 import cloud.glitchdev.rfu.model.party.FishingParty
 import cloud.glitchdev.rfu.utils.dsl.isIgnored
+import cloud.glitchdev.rfu.utils.network.PartyWebSocket
 import java.util.concurrent.CopyOnWriteArrayList
 
 object PartyFinderEvents {
@@ -152,7 +152,10 @@ object PartyFinderEvents {
     }
 
     fun handleCreated(party: FishingParty) {
-        _parties.removeIf { it.user == party.user }
+        val existed = _parties.removeIf { it.user == party.user }
+        if (existed) {
+            PartyWebSocket.syncParties()
+        }
         _parties.add(party)
         refreshParties()
         PartyCreated.runTasks(party)